backend/requirements.txt
# Flask Backend Requirements
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0
openai>=1.0.0
reportlab>=4.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Additional dependencies
Pillow>=10.0.0